Extract some common test helpers
