+++ src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Refactoring/RenameEndpointTest.cs
@@ -17,6 +17,7 @@ using Microsoft.AspNetCore.Razor.LanguageServer.Common;
 using Microsoft.AspNetCore.Razor.LanguageServer.EndpointContracts;
 using Microsoft.AspNetCore.Razor.LanguageServer.Extensions;
 using Microsoft.AspNetCore.Razor.LanguageServer.Protocol;
+using Microsoft.AspNetCore.Razor.LanguageServer.Test;
 using Microsoft.AspNetCore.Razor.LanguageServer.Test.Common;
 using Microsoft.AspNetCore.Razor.Test.Common;
 using Microsoft.CodeAnalysis.CSharp;
@@ -651,16 +652,6 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Refactoring.Test
             return endpoint;
         }
 
-        internal class TestProjectSnapshotManagerAccessor : ProjectSnapshotManagerAccessor
-        {
-            public TestProjectSnapshotManagerAccessor(ProjectSnapshotManagerBase instance)
-            {
-                Instance = instance;
-            }
-
-            public override ProjectSnapshotManagerBase Instance { get; }
-        }
-
         private class RenameLanguageServer : ClientNotifierServiceBase
         {
             private readonly CSharpTestLspServer _csharpServer;
@@ -701,29 +692,7 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Refactoring.Test
 
                 var result = await _csharpServer.ExecuteRequestAsync<RenameParams, WorkspaceEdit>(Methods.TextDocumentRenameName, renameRequest, CancellationToken.None);
 
-                return new ResponseRouterReturn(result);
-            }
-        }
-
-        private class ResponseRouterReturn : IResponseRouterReturns
-        {
-            private WorkspaceEdit _result;
-
-            public ResponseRouterReturn(WorkspaceEdit result)
-            {
-                _result = result;
-            }
-
-            public Task<TResponse> Returning<TResponse>(CancellationToken cancellationToken)
-            {
-                Assert.IsType<TResponse>(_result);
-
-                return Task.FromResult((TResponse)(object)_result);
-            }
-
-            public Task ReturningVoid(CancellationToken cancellationToken)
-            {
-                throw new NotImplementedException();
+                return new TestResponseRouterReturn(result);
             }
         }
     }
